feat: separate log dialog editing
